Rearrange namespaces, module structure and exports
