+++ metasync.js
@@ -1,21 +1,19 @@
 'use strict';
 
-const flow = require('./lib/flow');
-const arrayUtils = require('./lib/array.utils');
-const arrayChain = require('./lib/array.chain');
-const collectors = require('./lib/collectors');
-const queue = require('./lib/queue');
-const throttle = require('./lib/throttle');
+const api = {};
+api.metasync = {};
+module.exports = api.metasync;
 
-const metasync = {};
-module.exports = metasync;
+const submodules = [
+  'flow', // Flow control
+  'array', // Array utilities
+  'chain', // Process arrays sync and async array in chain
+  'collectors', // DataCollector and KeyCollector
+  'queue', // Concurrency
+  'throttle' // Throttling
+];
 
-Object.assign(
-  metasync,
-  flow, // Flow control
-  arrayUtils, // Array utilities
-  arrayChain, // Process arrays sync and async array in chain
-  collectors, // DataCollector and KeyCollector
-  queue, // Concurrency
-  throttle // Throttling
-);
+submodules
+  .map(path => './lib/' + path)
+  .map(require)
+  .map(exports => exports(api));
